Extract MetaCodec persistence codec from GovernancePipeline
Move the five pure (de)serialization methods (SessionMeta and
SessionStateSnapshot <-> JSON-able dict) plus the _decode_budget_dims helper
into a stateless MetaCodec collaborator. The pipeline constructs self._codec
and delegates the seven call sites. No store handle, no residency — just
structural mapping, so it is trivially testable and swappable.

Result types EscalationContext/TransformSuggestion are no longer re-exported
through pipeline; the one test relying on that now imports them from the
canonical tracemill.governance.results. No behavior change: 1884 passed / 4 skipped.
